Exztract UniqueId helper
