+++ dynamodb/dynamo_lock_test_utils.go
@@ -1,40 +1,18 @@
 package dynamodb
 
 import (
-	"bytes"
 	"fmt"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/gruntwork-io/terragrunt/aws_helper"
 	"github.com/gruntwork-io/terragrunt/options"
+	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/stretchr/testify/assert"
-	"math/rand"
 	"testing"
-	"time"
 )
 
 // For simplicity, do all testing in the us-east-1 region
 const DEFAULT_TEST_REGION = "us-east-1"
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
-}
-
-// Returns a unique (ish) id we can use to name resources so they don't conflict with each other. Uses base 62 to
-// generate a 6 character string that's unlikely to collide with the handful of tests we run in parallel. Based on code
-// here: http://stackoverflow.com/a/9543797/483528
-func uniqueId() string {
-	const BASE_62_CHARS = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
-	const UNIQUE_ID_LENGTH = 6 // Should be good for 62^6 = 56+ billion combinations
-
-	var out bytes.Buffer
-
-	for i := 0; i < UNIQUE_ID_LENGTH; i++ {
-		out.WriteByte(BASE_62_CHARS[rand.Intn(len(BASE_62_CHARS))])
-	}
-
-	return out.String()
-}
 
 // Create a DynamoDB client we can use at test time. If there are any errors creating the client, fail the test.
 func createDynamoDbClientForTest(t *testing.T) *dynamodb.DynamoDB {
@@ -55,7 +33,7 @@ func createDynamoDbClientForTest(t *testing.T) *dynamodb.DynamoDB {
 }
 
 func uniqueTableNameForTest() string {
-	return fmt.Sprintf("terragrunt_test_%s", uniqueId())
+	return fmt.Sprintf("terragrunt_test_%s", util.UniqueId())
 }
 
 func cleanupTableForTest(t *testing.T, tableName string, client *dynamodb.DynamoDB) {
@@ -64,7 +42,7 @@ func cleanupTableForTest(t *testing.T, tableName string, client *dynamodb.Dynamo
 }
 
 func assertCanWriteToTable(t *testing.T, tableName string, client *dynamodb.DynamoDB) {
-	item := createKeyFromItemId(uniqueId())
+	item := createKeyFromItemId(util.UniqueId())
 
 	_, err := client.PutItem(&dynamodb.PutItemInput{
 		TableName: aws.String(tableName),
