+++ util/random.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"bytes"
 	"math/rand"
 	"time"
 )
@@ -39,4 +40,24 @@ func GetRandomTime(lowerBound, upperBound time.Duration) time.Duration {
 func random(min int, max int) int {
 	rand.Seed(time.Now().UnixNano())
 	return rand.Intn(max-min) + min
+}
+
+func init() {
+	rand.Seed(time.Now().UnixNano())
+}
+
+const BASE_62_CHARS = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+const UNIQUE_ID_LENGTH = 6 // Should be good for 62^6 = 56+ billion combinations
+
+// Returns a unique (ish) id we can use to name resources so they don't conflict with each other. Uses base 62 to
+// generate a 6 character string that's unlikely to collide with the handful of tests we run in parallel. Based on code
+// here: http://stackoverflow.com/a/9543797/483528
+func UniqueId() string {
+	var out bytes.Buffer
+
+	for i := 0; i < UNIQUE_ID_LENGTH; i++ {
+		out.WriteByte(BASE_62_CHARS[rand.Intn(len(BASE_62_CHARS))])
+	}
+
+	return out.String()
 }
